refactor: 포인트 사용 로직 수정
