+++ apps/commerce-api/src/main/java/com/loopers/application/order/CouponUseService.java
@@ -1,21 +1,16 @@
-package com.loopers.application.userCoupon;
+package com.loopers.application.order;
 
 import com.loopers.domain.coupon.Coupon;
 import com.loopers.domain.coupon.CouponService;
 import com.loopers.domain.userCoupon.UserCoupon;
 import com.loopers.domain.userCoupon.UserCouponService;
-import com.loopers.support.error.CoreException;
-import com.loopers.support.error.ErrorType;
 import lombok.RequiredArgsConstructor;
-import org.springframework.orm.ObjectOptimisticLockingFailureException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.loopers.support.utils.Validation.Message.MESSAGE_COUPON_UNUSABLE;
-
 @Service
 @RequiredArgsConstructor
-public class UserCouponUseService {
+public class CouponUseService {
 
     private final CouponService couponService;
     private final UserCouponService userCouponService;
@@ -28,11 +23,7 @@ public class UserCouponUseService {
         couponService.validateAmount(coupon, orderAmount);
         int discountAmount = couponService.calculateDiscountAmount(coupon, orderAmount);
 
-        try {
-            userCouponService.use(userCouponId, userId);
-        } catch (ObjectOptimisticLockingFailureException e) {
-            throw new CoreException(ErrorType.CONFLICT, MESSAGE_COUPON_UNUSABLE);
-        }
+        userCouponService.use(userCouponId, userId);
 
         return discountAmount;
     }
