+++ apps/commerce-api/src/test/java/com/loopers/application/order/CouponUseServiceConcurrencyTest.java
@@ -1,4 +1,4 @@
-package com.loopers.application.userCoupon;
+package com.loopers.application.order;
 
 import com.loopers.domain.coupon.Coupon;
 import com.loopers.domain.userCoupon.CouponUsageHistory;
@@ -7,6 +7,8 @@ import com.loopers.domain.userCoupon.UserCouponStatus;
 import com.loopers.infrastructure.coupon.CouponJpaRepository;
 import com.loopers.infrastructure.userCoupon.CouponUsageHistoryJpaRepository;
 import com.loopers.infrastructure.userCoupon.UserCouponJpaRepository;
+import com.loopers.utils.DatabaseCleanUp;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +24,11 @@ import static com.loopers.domain.coupon.DiscountType.PRICE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
-class UserCouponUseServiceConcurrencyTest {
+class CouponUseServiceConcurrencyTest {
+    // sut --
+    @Autowired
+    private CouponUseService couponUseService;
+
     // orm--
     @Autowired
     private UserCouponJpaRepository userCouponJpaRepository;
@@ -30,14 +36,17 @@ class UserCouponUseServiceConcurrencyTest {
     private CouponJpaRepository couponJpaRepository;
     @Autowired
     private CouponUsageHistoryJpaRepository couponUsageHistoryJpaRepository;
-
-    // sut --
     @Autowired
-    private UserCouponUseService userCouponUseService;
+    private DatabaseCleanUp databaseCleanUp;
+
+    @AfterEach
+    void tearDown() {
+        databaseCleanUp.truncateAllTables();
+    }
 
     @DisplayName("낙관적 락 적용 시, 동시에 쿠폰을 사용해도 한 번만 사용된다.")
     @Test
-    void getDetailWithLock() throws InterruptedException {
+    void useCouponWithOptimisticLock() throws InterruptedException {
         // arrange
         Coupon coupon = couponJpaRepository.save(new Coupon("3천원 할인", 100, 0, PRICE, 3000, null, 10000));
         UserCoupon userCoupon = userCouponJpaRepository.save(UserCoupon.create(coupon.getId(), 1L));
@@ -52,10 +61,10 @@ class UserCouponUseServiceConcurrencyTest {
         for (int i = 0; i < threadCount; i++) {
             executor.submit(() -> {
                 try {
-                    userCouponUseService.use(userCoupon.getId(), userCoupon.getUserId(), 10000);
+                    couponUseService.use(userCoupon.getId(), userCoupon.getUserId(), 10000);
                     successCount.incrementAndGet();
                 } catch (Exception e) {
-                    System.out.println("[FAIL]: " + e.getMessage());
+                    System.out.println("[FAIL]: "+e.getMessage());
                     failCount.incrementAndGet();
                 } finally {
                     latch.countDown();
@@ -67,8 +76,9 @@ class UserCouponUseServiceConcurrencyTest {
         executor.shutdown();
 
         // assert
+        System.out.println("[SUCCESS COUNT]: "+successCount.get());
+        System.out.println("[FAIL COUNT]: "+failCount.get());
         assertThat(successCount.get()).isEqualTo(1);
-        assertThat(failCount.get()).isEqualTo(threadCount-successCount.get());
 
         UserCoupon result = userCouponJpaRepository.findById(userCoupon.getId()).orElseThrow();
         assertThat(result.getStatus()).isEqualTo(UserCouponStatus.USED);
