+++ apps/commerce-api/src/test/java/com/loopers/domain/userCoupon/UserCouponServiceDomainTest.java
@@ -135,10 +135,9 @@ class UserCouponServiceDomainTest {
             given(userCouponRepository.saveHistory(couponUsageHistory)).willReturn(couponUsageHistory);
 
             // when
-            CouponUsageHistory result = userCouponService.saveHistory(couponUsageHistory);
+            userCouponService.saveHistory(couponUsageHistory);
 
             // then
-            assertThat(result).isEqualTo(couponUsageHistory);
             verify(userCouponRepository, times(1)).saveHistory(couponUsageHistory);
         }
     }
